refactor(cli): move all Metro imports over from `metro(-*)/src/` to `metro(-*)/private/`
