+++ packages/@expo/cli/src/start/server/metro/TerminalReporter.ts
@@ -1,8 +1,8 @@
 // This file represents an abstraction on the metro TerminalReporter.
 // We use this abstraction to safely extend the TerminalReporter for our own custom logging.
 import chalk from 'chalk';
-import UpstreamTerminalReporter from 'metro/src/lib/TerminalReporter';
-import { Terminal } from 'metro-core';
+import UpstreamTerminalReporter from 'metro/private/lib/TerminalReporter';
+import type { Terminal } from 'metro-core';
 import type { WatcherStatus } from 'metro-file-map';
 import util from 'util';
