+++ packages/@expo/cli/src/start/server/metro/instantiateMetro.ts
@@ -5,12 +5,12 @@ import chalk from 'chalk';
 import http from 'http';
 import type Metro from 'metro';
 import { ReadOnlyGraph } from 'metro';
-import Bundler from 'metro/src/Bundler';
-import type { TransformOptions } from 'metro/src/DeltaBundler/Worker';
-import MetroHmrServer from 'metro/src/HmrServer';
-import RevisionNotFoundError from 'metro/src/IncrementalBundler/RevisionNotFoundError';
-import formatBundlingError from 'metro/src/lib/formatBundlingError';
-import { loadConfig, resolveConfig, ConfigT } from 'metro-config';
+import type Bundler from 'metro/private/Bundler';
+import type { TransformOptions } from 'metro/private/DeltaBundler/Worker';
+import type MetroHmrServer from 'metro/private/HmrServer';
+import RevisionNotFoundError from 'metro/private/IncrementalBundler/RevisionNotFoundError';
+import formatBundlingError from 'metro/private/lib/formatBundlingError';
+import { loadConfig, resolveConfig, type ConfigT } from 'metro-config';
 import { Terminal } from 'metro-core';
 import util from 'node:util';
 import path from 'path';
