+++ packages/@expo/cli/src/start/server/metro/metroVirtualModules.ts
@@ -4,9 +4,9 @@
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
  */
-import Bundler from 'metro/src/Bundler';
-import DependencyGraph from 'metro/src/node-haste/DependencyGraph';
-import { FileSystem } from 'metro-file-map';
+import type Bundler from 'metro/private/Bundler';
+import type DependencyGraph from 'metro/private/node-haste/DependencyGraph';
+import type { FileSystem } from 'metro-file-map';
 
 type ExpoPatchedFileSystem = Omit<FileSystem, 'getSha1'> & {
   getSha1: FileSystem['getSha1'] & { __patched?: boolean };
